Rename message componenets and add comments to them
